moved svg object url logic back to bhce
